[ONL-7918] rename sort-direction.ts to index.ts
