+++ src/components/ec-smart-table/ec-smart-table.vue
@@ -123,7 +123,7 @@ import {
 
 import useEcPagination from '../../composables/use-ec-pagination';
 import useEcSorting from '../../composables/use-ec-sorting';
-import { SortDirection } from '../../enums/sort-direction';
+import { SortDirection } from '../../enums';
 import * as SortDirectionCycle from '../../enums/sort-direction-cycle';
 import EcLoading from '../ec-loading';
 import EcSmartTableHeading from '../ec-smart-table-heading';
@@ -225,3 +225,4 @@ function getEcTableSlots() {
   return tableSlots;
 }
 </script>
+../../enums
