+++ src/components/ec-table-head/ec-table-head.spec.js
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils';
 
 import { withMockedConsole } from '../../../tests/utils/console';
-import { SortDirection } from '../../enums/sort-direction';
+import { SortDirection } from '../../enums';
 import EcTableHead from './ec-table-head.vue';
 
 describe('EcTableHead', () => {
