+++ src/enums/sort-direction-cycle.js
@@ -1,4 +1,4 @@
-import { SortDirection } from './sort-direction';
+import { SortDirection } from '.';
 
 export const LOWEST_FIRST = [SortDirection.ASC, SortDirection.DESC];
 export const HIGHEST_FIRST = [SortDirection.DESC, SortDirection.ASC];
