+++ src/main.js
@@ -1,4 +1,4 @@
-import { SortDirection } from './enums/sort-direction';
+import { SortDirection } from './enums';
 import * as SortDirectionCycle from './enums/sort-direction-cycle';
 import * as TooltipPlacement from './enums/tooltip-placement';
